Update for generating QdQ

+++ examples/keras/image_recognition/imagenet_prepare/build_imagenet_data.py
@@ -377,7 +377,7 @@ def _process_image_files(name, filenames, synsets, labels, humans, num_shards):
   assert len(filenames) == len(humans)
 
   # Break all images into batches with a [ranges[i][0], ranges[i][1]].
-  spacing = np.linspace(0, len(filenames), FLAGS.num_threads + 1).astype(np.int)
+  spacing = np.linspace(0, len(filenames), FLAGS.num_threads + 1).astype(np.int32)
   ranges = []
   threads = []
   for i in xrange(len(spacing) - 1):
